主题：完善STM32客户端
详细信息：
1、解除J-link调试端口的占用，增加GPIO使用资源；
2、PC13作为系统运行指示灯；
3、添加MQ-2传感器模块，用于监测烟雾和甲烷（天然气）

+++ 智能家居/hardware/light_sensor.h
@@ -3,10 +3,11 @@
 
 #include "stm32f10x.h"
 #include "OLED.h"
+#include "delay.h"
 
 void light_sensor_init(void);
 
-uint16_t get_light_sensor_value(void);
+float get_light_sensor_value(void);
 
 float get_light_sensor_voltage_value(void);
